Update software.md
apparently you have to force the command. I also noticed that the command "npm install --force -g aws-cdk@latest" has to be executed at-least one time irrespective of which lab you are doing. 
This will fix things temporarily but if there is a way to upgrade the version of cdk ( even without bothering the person who is doing the labs), then thats the best way.