update prerequisites (#46)
* update prerequisites

Include latest version of cdk to avoid errors

Without the update the following error is being thrown :- 
This CDK CLI is not compatible with the CDK library used by your application. Please upgrade the CLI to the latest version.
(Cloud assembly schema version mismatch: Maximum schema version supported is 17.0.0, but found 18.0.0)

* Update software.md

apparently you have to force the command. I also noticed that the command "npm install --force -g aws-cdk@latest" has to be executed at-least one time irrespective of which lab you are doing. 
This will fix things temporarily but if there is a way to upgrade the version of cdk ( even without bothering the person who is doing the labs), then thats the best way.

* Update aws-cdk version

After giving it some thought,  realized that not all the latest versions might be compatible. aws-cdk@2.25.0 is stable and the lab works well with this version.